refactor(routes): move API endpoints into dedicated Express router (#31)
